+++ packages/cozy-procedures/babel.config.js
@@ -1,4 +1,15 @@
 module.exports = {
   presets: ['cozy-app'],
-  plugins: ['inline-react-svg']
+  ignore: ['*.spec.js', '*.spec.jsx'],
+  env: {
+    transpilation: {
+      presets: [
+        [
+          'cozy-app',
+          { presetEnv: { modules: false }, transformRuntime: { helpers: true } }
+        ]
+      ],
+      plugins: ['inline-react-svg']
+    }
+  }
 }