+++ packages/cozy-procedures/src/containers/Amount.jsx
@@ -1,19 +1,15 @@
-import { connect } from 'react-redux'
-import context from '../redux/context'
+import connectWithProcedures from '../redux/connectWithProcedures'
 import Amount from '../components/Amount'
-import { getSlice, update } from '../redux/amountSlice'
+import { getAmount, updateAmount } from '../redux/procedureDataSlice'
 
 const mapStateToProps = state => ({
-  amount: getSlice(state)
+  amount: getAmount(state)
 })
 
-const mapDispatchToProps = dispatch => ({
-  updateAmount: value => dispatch(update(value))
-})
+const mapDispatchToProps = {
+  updateAmount
+}
 
-export default connect(
-  mapStateToProps,
-  mapDispatchToProps,
-  null,
-  { context }
-)(Amount)
+export default connectWithProcedures(mapStateToProps, mapDispatchToProps)(
+  Amount
+)