+++ packages/cozy-procedures/src/containers/DocumentsDataForm.jsx
@@ -1,5 +1,4 @@
-import { connect } from 'react-redux'
-import context from '../redux/context'
+import connectWithProcedures from '../redux/connectWithProcedures'
 import {
   fetchDocumentsByCategory,
   getFiles,
@@ -7,7 +6,8 @@ import {
   linkDocumentSuccess,
   getFilesStatus,
   setDocumentLoading,
-  fetchDocumentError
+  fetchDocumentError,
+  setLoadingFalse
 } from '../redux/documentsDataSlice'
 
 const mapStateToProps = state => ({
@@ -20,15 +20,11 @@ const mapDispatchToProps = {
   unlinkDocument,
   linkDocumentSuccess,
   setDocumentLoading,
-  fetchDocumentError
+  fetchDocumentError,
+  setLoadingFalse
 }
 
 const DocumentsDataFormContainer = Component =>
-  connect(
-    mapStateToProps,
-    mapDispatchToProps,
-    null,
-    { context }
-  )(Component)
+  connectWithProcedures(mapStateToProps, mapDispatchToProps)(Component)
 
 export default DocumentsDataFormContainer