+++ packages/cozy-procedures/src/containers/Duration.jsx
@@ -1,21 +1,20 @@
-import { connect } from 'react-redux'
-import context from '../redux/context'
+import connectWithProcedures from '../redux/connectWithProcedures'
 import Duration from '../components/Duration'
-import { getSlice as getDuration, update } from '../redux/durationSlice'
-import { getSlice as getAmount } from '../redux/amountSlice'
+import {
+  getAmount,
+  getDuration,
+  updateDuration
+} from '../redux/procedureDataSlice'
 
 const mapStateToProps = state => ({
   duration: getDuration(state),
   amount: getAmount(state)
 })
 
-const mapDispatchToProps = dispatch => ({
-  updateDuration: value => dispatch(update(value))
-})
+const mapDispatchToProps = {
+  updateDuration
+}
 
-export default connect(
-  mapStateToProps,
-  mapDispatchToProps,
-  null,
-  { context }
-)(Duration)
+export default connectWithProcedures(mapStateToProps, mapDispatchToProps)(
+  Duration
+)