+++ packages/cozy-procedures/src/containers/Procedure.jsx
@@ -1,6 +1,8 @@
-import { connect } from 'react-redux'
-
-import Context from '../redux/context'
+import connectWithProcedures from '../redux/connectWithProcedures'
+import {
+  getInitialized,
+  initializationSuccess
+} from '../redux/procedureDataSlice'
 import {
   init as initPersonalData,
   fetchMyself,
@@ -8,31 +10,25 @@ import {
 } from '../redux/personalDataSlice'
 import {
   init as initDocuments,
-  fetchDocumentsByCategory,
-  setProcedureStatus,
-  getInitiated
+  fetchDocumentsByCategory
 } from '../redux/documentsDataSlice'
 
 import Procedure from '../Procedure'
 import withLocales from '../withLocales'
 
 const mapStateToProps = state => ({
-  initiated: getInitiated(state)
+  initialized: getInitialized(state)
 })
+
 const mapDispatchToProps = {
   initPersonalData,
   fetchMyself,
   initDocuments,
   fetchDocumentsByCategory,
-  setProcedureStatus,
-  fetchBankAccountsStats
+  fetchBankAccountsStats,
+  initializationSuccess
 }
 
 export default withLocales(
-  connect(
-    mapStateToProps,
-    mapDispatchToProps,
-    null,
-    { context: Context }
-  )(Procedure)
+  connectWithProcedures(mapStateToProps, mapDispatchToProps)(Procedure)
 )