+++ packages/cozy-doctypes/CHANGELOG.md
@@ -3,6 +3,47 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+# [1.56.0](https://github.com/cozy/cozy-libs/compare/cozy-doctypes@1.55.2...cozy-doctypes@1.56.0) (2019-07-18)
+
+
+### Features
+
+* **cozy-doctypes:** Add fixed charges to BankAccountStats::sum ([b37cf33](https://github.com/cozy/cozy-libs/commit/b37cf33))
+
+
+
+
+
+## [1.55.2](https://github.com/cozy/cozy-libs/compare/cozy-doctypes@1.55.1...cozy-doctypes@1.55.2) (2019-07-18)
+
+**Note:** Version bump only for package cozy-doctypes
+
+
+
+
+
+## [1.55.1](https://github.com/cozy/cozy-libs/compare/cozy-doctypes@1.55.0...cozy-doctypes@1.55.1) (2019-07-17)
+
+
+### Bug Fixes
+
+* **cozy-doctypes:** Fix CozyFiles overrideFileForPath ([4329c4f](https://github.com/cozy/cozy-libs/commit/4329c4f))
+
+
+
+
+
+# [1.55.0](https://github.com/cozy/cozy-libs/compare/cozy-doctypes@1.54.0...cozy-doctypes@1.55.0) (2019-07-16)
+
+
+### Features
+
+* Can override a file with the same name ([ff1d440](https://github.com/cozy/cozy-libs/commit/ff1d440))
+
+
+
+
+
 # [1.54.0](https://github.com/cozy/cozy-libs/compare/cozy-doctypes@1.53.2...cozy-doctypes@1.54.0) (2019-07-11)